#649 More webserver refactoring
